Added Configuration
DataSourceConfig is for configuring the Postgres database to add the CITEXT type extension, more info in the javadoc of the config class.

I still haven't figured out how to make Spring JPA instantiate the fields with CITEXT when it's a Postgres Database, and with a regular string type otherwise. (I am however able to tell whether I'm working with a Postgres DB or another DB, checkout the DataSourceConfig class for that)