Added Security
New Security package with SecurityConfiguration for configuration and UserDetailsServiceImpl for getting users and their roles from the database.

Also Made updates to the UserDao and UserService.

The Launcher class (the one that has the main method) now instantiates a PasswordEncryption Bean mainly for the UserService to use it for encrypting passwords.